# specs/001-constitution-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Constitution API
  description: API for managing constitution principles, performance standards, and governance
  version: 1.0.0
  contact:
    name: Constitution API Team
    email: api@constitution.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.constitution.com
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login and get JWT tokens
      description: Authenticate user and return access and refresh tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: password123
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: JWT access token
                    example: eyJhbGciOiJIUzI1NiIs...
                  refresh_token:
                    type: string
                    description: JWT refresh token
                    example: eyJhbGciOiJIUzI1NiIs...
                  expires_in:
                    type: integer
                    description: Token expiration time in seconds
                    example: 900
        '401':
          description: Invalid credentials
        '422':
          description: Validation error
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      description: Get new access token using refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - refresh_token
              properties:
                refresh_token:
                  type: string
                  description: Valid refresh token
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    description: New JWT access token
                  expires_in:
                    type: integer
                    description: Token expiration time in seconds
        '401':
          description: Invalid refresh token

  # Principles Endpoints
  /principles:
    get:
      tags:
        - Principles
      summary: Get all constitution principles
      description: Retrieve all active constitution principles with pagination
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: category
          in: query
          schema:
            type: string
        - name: priority
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 5
      responses:
        '200':
          description: List of principles
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Principle'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '400':
          description: Invalid query parameters

    post:
      tags:
        - Principles
      summary: Create a new principle
      description: Add a new constitution principle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePrinciple'
      responses:
        '201':
          description: Principle created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principle'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /principles/{id}:
    get:
      tags:
        - Principles
      summary: Get a specific principle
      description: Retrieve a single principle by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Principle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principle'
        '404':
          description: Principle not found

    put:
      tags:
        - Principles
      summary: Update a principle
      description: Update an existing principle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePrinciple'
      responses:
        '200':
          description: Principle updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principle'
        '400':
          description: Invalid input
        '404':
          description: Principle not found

    delete:
      tags:
        - Principles
      summary: Delete a principle
      description: Mark a principle as inactive (soft delete)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Principle deleted successfully
        '404':
          description: Principle not found
  /principles/search:
    get:
      tags:
        - Principles
      summary: Search principles
      description: Full-text search across principles
      parameters:
        - name: query
          in: query
          required: true
          schema:
            type: string
          description: Search query string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Principle'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
                  total:
                    type: integer
                    description: Total number of results

  # Performance Standards Endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Principle:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
          minimum: 1
          maximum: 5
        metadata:
          type: object
          properties:
            category:
              type: string
            tags:
              type: array
              items:
                type: string
            relatedPrinciples:
              type: array
              items:
                type: string
            examples:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [code, diagram, text]
                  title:
                    type: string
                  content:
                    type: string
                  language:
                    type: string
        currentVersion:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        performanceStandards:
          type: array
          items:
            $ref: '#/components/schemas/PerformanceStandard'
        loadTestRequirements:
          type: array
          items:
            $ref: '#/components/schemas/LoadTestRequirement'
        qualityGates:
          type: array
          items:
            $ref: '#/components/schemas/QualityGate'
    CreatePrinciple:
      type: object
      required:
        - title
        - description
        - priority
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          minLength: 10
        priority:
          type: integer
          minimum: 1
          maximum: 5
        metadata:
          type: object
    UpdatePrinciple:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          minLength: 10
        priority:
          type: integer
          minimum: 1
          maximum: 5
        metadata:
          type: object
    PerformanceStandard:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        endpointType:
          type: string
        targetResponseTime:
          type: integer
        metricType:
          type: string
          enum: [95th_percentile, average, max, min]
        concurrentUsers:
          type: integer
        additionalMetrics:
          type: object
          properties:
            cpuUsage:
              type: number
              maximum: 100
            memoryUsage:
              type: number
              maximum: 100
            errorRate:
              type: number
              maximum: 100
        description:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    LoadTestRequirement:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        testType:
          type: string
          enum: [smoke, load, stress, spike]
        endpointPattern:
          type: string
        testParameters:
          type: object
          properties:
            rampUp:
              type: object
              properties:
                from:
                  type: integer
                to:
                  type: integer
                duration:
                  type: string
            constantLoad:
              type: object
              properties:
                users:
                  type: integer
                duration:
                  type: string
            stress:
              type: object
              properties:
                load:
                  type: string
                duration:
                  type: string
        expectedRate:
          type: integer
        maxResponseTime:
          type: integer
        maxErrorRate:
          type: number
        scriptPath:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    QualityGate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          type: string
          enum: [automated, manual]
        tool:
          type: string
        requirement:
          type: string
        criteria:
          type: object
          properties:
            threshold:
              type: number
            unit:
              type: string
            conditions:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  operator:
                    type: string
                  value:
                    type: any
        isRequired:
          type: boolean
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...